requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
sentence-transformers==3.1.1
numpy<2
transformers==4.41.2
chromadb==0.5.5
psycopg2-binary==2.9.9
pgvector==0.2.5
google-generativeai==0.3.0

# Use CPU-only PyTorch wheels to avoid pulling CUDA/NVIDIA packages.
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu